# docs/swagger.yaml
basePath: /api
definitions:
...
  preference.OptionItem:
    properties:
      label:
        description: 显示名称
        type: string
      value:
        description: 标签值
        type: string
    type: object
  preference.OptionItemWithSelected:
    properties:
      label:
        description: 显示名称
        type: string
      selected:
        description: 是否已选择
        type: boolean
      value:
        description: 标签值
        type: string
    type: object
  preference.OptionsResponse:
    properties:
      allergies:
        description: 过敏选项
        items:
          $ref: '#/definitions/preference.OptionItem'
        type: array
      flavors:
        description: 口味选项
        items:
          $ref: '#/definitions/preference.OptionItem'
        type: array
      taboos:
        description: 忌口选项
        items:
          $ref: '#/definitions/preference.OptionItem'
        type: array
    type: object
  preference.UpdatePreferencesRequest:
    properties:
      allergies:
        description: 过敏选项值列表
        items:
          type: string
        type: array
      flavors:
        description: 口味选项值列表
        items:
          type: string
        type: array
      taboos:
        description: 忌口选项值列表
        items:
          type: string
        type: array
    type: object
  preference.UserPreferencesResponse:
    properties:
      allergies:
        description: 过敏偏好
        items:
          $ref: '#/definitions/preference.OptionItemWithSelected'
        type: array
      flavors:
        description: 口味偏好
        items:
          $ref: '#/definitions/preference.OptionItemWithSelected'
        type: array
      taboos:
        description: 忌口偏好
        items:
          $ref: '#/definitions/preference.OptionItemWithSelected'
        type: array
    type: object
...
  recipe.IngredientRequest:
    properties:
      name:
        type: string
      notes:
        type: string
      quantity:
        type: number
      text_quantity:
        type: string
      unit:
        type: string
    required:
    - name
    - text_quantity
    type: object
  recipe.IngredientResponse:
    properties:
      category:
        type: string
      id:
        type: integer
      name:
        type: string
      notes:
        type: string
      quantity:
        type: number
      text_quantity:
        type: string
      unit:
        type: string
    type: object
  recipe.RecipeCreateRequest:
    properties:
      additional_notes:
        items:
          type: string
        type: array
      category:
        type: string
      cook_time_minutes:
        type: integer
      description:
        type: string
      difficulty:
        type: integer
      id:
        type: string
      images:
        items:
          type: string
        type: array
      ingredients:
        items:
          $ref: '#/definitions/recipe.IngredientRequest'
        type: array
      name:
        type: string
      prep_time_minutes:
        type: integer
      servings:
        type: integer
      steps:
        items:
          $ref: '#/definitions/recipe.StepRequest'
        type: array
      tags:
        items:
          type: string
        type: array
      total_time_minutes:
        type: integer
    required:
    - category
    - difficulty
    - name
    - servings
    type: object
  recipe.RecipeResponse:
    properties:
      additional_notes:
        items:
          type: string
        type: array
      category:
        type: string
      cook_time_minutes:
        type: integer
      description:
        type: string
      difficulty:
        type: integer
      id:
        type: string
      image_path:
        type: string
      images:
        items:
          type: string
        type: array
      ingredients:
        items:
          $ref: '#/definitions/recipe.IngredientResponse'
        type: array
      name:
        type: string
      prep_time_minutes:
        type: integer
      servings:
        type: integer
      steps:
        items:
          $ref: '#/definitions/recipe.StepResponse'
        type: array
      tags:
        $ref: '#/definitions/types.TagsGrouped'
      total_time_minutes:
        type: integer
    type: object
  recipe.RecipeUpdateRequest:
    properties:
      additional_notes:
        items:
          type: string
        type: array
      category:
        type: string
      cook_time_minutes:
        type: integer
      description:
        type: string
      difficulty:
        type: integer
      images:
        items:
          type: string
        type: array
      ingredients:
        items:
          $ref: '#/definitions/recipe.IngredientRequest'
        type: array
      name:
        type: string
      prep_time_minutes:
        type: integer
      servings:
        type: integer
      steps:
        items:
          $ref: '#/definitions/recipe.StepRequest'
        type: array
      tags:
        items:
          type: string
        type: array
      total_time_minutes:
        type: integer
    type: object
  recipe.StepRequest:
    properties:
      description:
        type: string
      step:
        type: integer
    required:
    - description
    - step
    type: object
  recipe.StepResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      step:
        type: integer
    type: object
...
  types.RecipeListItem:
    properties:
      category:
        type: string
      description:
        type: string
      difficulty:
        type: integer
      id:
        type: string
      image_path:
        type: string
      name:
        type: string
      tags:
        $ref: '#/definitions/types.TagsGrouped'
      total_time_minutes:
        type: integer
    type: object
  types.TagsGrouped:
    properties:
      cuisines:
        items:
          type: string
        type: array
      flavors:
        items:
          type: string
        type: array
      scenes:
        items:
          type: string
        type: array
    type: object
...
host: localhost:18000
info:
  contact: {}
  description: 菜谱管理系统后端 API
  title: Choosy API
  version: "1.0"
paths:
...
  /api/preferences:
    get:
      description: 获取口味、忌口、过敏的所有可选选项（从缓存索引获取，性能最优）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/preference.OptionsResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取所有偏好选项
      tags:
      - preference
  /api/recipes:
    get:
      parameters:
      - description: 分类
        in: query
        name: category
        type: string
      - description: 搜索关键词
        in: query
        name: search
        type: string
      - default: 100
        description: 限制数量
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.RecipeListItem'
            type: array
      summary: 获取菜谱列表
      tags:
      - recipes
    post:
      consumes:
      - application/json
      parameters:
      - description: 菜谱信息
        in: body
        name: recipe
        required: true
        schema:
          $ref: '#/definitions/recipe.RecipeCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/recipe.RecipeResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 创建新菜谱
      tags:
      - recipes
  /api/recipes/{recipe_id}:
    delete:
      parameters:
      - description: 菜谱ID
        in: path
        name: recipe_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 删除菜谱
      tags:
      - recipes
    get:
      parameters:
      - description: 菜谱ID
        in: path
        name: recipe_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/recipe.RecipeResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取菜谱详情
      tags:
      - recipes
    put:
      consumes:
      - application/json
      parameters:
      - description: 菜谱ID
        in: path
        name: recipe_id
        required: true
        type: string
      - description: 更新内容
        in: body
        name: recipe
        required: true
        schema:
          $ref: '#/definitions/recipe.RecipeUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/recipe.RecipeResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 更新菜谱
      tags:
      - recipes
  /api/recipes/batch:
    post:
      consumes:
      - application/json
      parameters:
      - description: 菜谱列表
        in: body
        name: recipes
        required: true
        schema:
          items:
            $ref: '#/definitions/recipe.RecipeCreateRequest'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/recipe.RecipeResponse'
            type: array
      summary: 批量创建菜谱
      tags:
      - recipes
...
  /api/user/preference:
    get:
      description: 获取当前用户的偏好设置（包含选中状态）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/preference.UserPreferencesResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取用户偏好
      tags:
      - preference
    put:
      consumes:
      - application/json
      description: 更新当前用户的偏好设置（全量替换）
      parameters:
      - description: 偏好设置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/preference.UpdatePreferencesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新用户偏好
      tags:
      - preference
...
securityDefinitions:
  Bearer:
    description: 输入 "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"